Master:
```
contour
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "testdata" => BenchmarkTools.Trial: 
	  memory estimate:  2.22 MiB
	  allocs estimate:  21188
	  --------------
	  minimum time:     2.405 ms (0.00% GC)
	  median time:      2.667 ms (0.00% GC)
	  mean time:        2.779 ms (3.18% GC)
	  maximum time:     6.316 ms (28.09% GC)
	  --------------
	  samples:          1794
	  evals/sample:     1

```

This branch:
```
contour
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "testdata" => BenchmarkTools.Trial: 
	  memory estimate:  1013.42 KiB
	  allocs estimate:  1189
	  --------------
	  minimum time:     1.828 ms (0.00% GC)
	  median time:      2.052 ms (0.00% GC)
	  mean time:        2.139 ms (1.24% GC)
	  maximum time:     5.281 ms (0.00% GC)
	  --------------
	  samples:          2329
	  evals/sample:     1
```

I need to add comments and squash.