@sjkelly 2020-01-28
With bit packing and removing tuples for ambiguous cases:
```
contour
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "testdata" => BenchmarkTools.Trial: 
	  memory estimate:  759.28 KiB
	  allocs estimate:  1189
	  --------------
	  minimum time:     1.471 ms (0.00% GC)
	  median time:      1.636 ms (0.00% GC)
	  mean time:        1.666 ms (1.60% GC)
	  maximum time:     3.942 ms (0.00% GC)
	  --------------
	  samples:          2992
	  evals/sample:     1
```