@sjkelly 2020-02-14
Closing in favor on #50